Make config object a property on Aptrac.
Move config from a modul with global state to a property in an Aptrac instance.
The main module now return the Aptrac constructor instead of an instance, so that the construction can use options for the config and database to use.
I had to refactor all tests and all commands to make this change